Refactor:#179 - move API mapper functions to src/api/mapper
