+++ src/api/calendarApi.ts
@@ -1,6 +1,6 @@
 import { getRequest } from '@/api';
+import { mapApiResponseToCalendar } from '@/api/mapper/calendarMapper';
 import { ApiCalendarResponse, CalendarResponse } from '@/interfaces/calendar';
-import { mapApiResponseToCalendar } from '@/lib/calendarMapper';
 
 export const getCalendarData = async (userId: number, date: string): Promise<CalendarResponse> => {
   const apiResponse: ApiCalendarResponse = await getRequest('/goals/todos/due-monthly', {
